Move config to metadata state
