+++ rs/execution_environment/src/query_handler.rs
@@ -329,8 +329,7 @@ impl InternalHttpQueryHandler {
 
         // Letting the canister grow arbitrarily when executing the
         // query is fine as we do not persist state modifications.
-        let subnet_available_memory =
-            full_subnet_memory_capacity(&self.config, state.get_ref().resource_limits());
+        let subnet_available_memory = full_subnet_memory_capacity(&self.config, state.get_ref());
         // Letting the canister use the full subnet memory reservation
         // is fine as we do not persist state modifications.
         let subnet_memory_reservation = self.config.subnet_memory_reservation;
